check at start

+++ src/io/kurumi/ntt/Launcher.java
@@ -9,15 +9,18 @@ import io.kurumi.ntt.model.*;
 import java.io.*;
 import java.util.*;
 import io.kurumi.ntt.twitter.track.FollowerTrackTask;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 public class Launcher extends BotFragment implements Thread.UncaughtExceptionHandler {
 
     public static final Launcher INSTANCE = new Launcher();
 
+    public static AtomicBoolean initIng = new AtomicBoolean(false);
+
     public Launcher() {
 
         addFragment(Backup.INSTANCE);
-        
+
         addFragment(GroupRepeat.INSTANCE);
 
 		addFragment(TwitterUI.INSTANCE);
@@ -26,10 +29,12 @@ public class Launcher extends BotFragment implements Thread.UncaughtExceptionHan
 		//addFragment(LuaEnv.INSTANCE.LuaFragmentOriginInstance);
 
         addFragment(TwitterArchive.INSTANCE);
-        
+
         addFragment(TwitterTrack.INSTANCE);
-        
+
         FollowerTrackTask.start();
+
+        new InitTask().start();
 
     }
 
@@ -38,7 +43,7 @@ public class Launcher extends BotFragment implements Thread.UncaughtExceptionHan
         Thread.setDefaultUncaughtExceptionHandler(INSTANCE);
 
 		TimeZone.setDefault(TimeZone.getTimeZone("Asia/Shanghai"));
-		
+
         /*
 
 		 BotServer.INSTACNCE.fragments.add(TGWebHookF.INSTANCE);
@@ -59,8 +64,6 @@ public class Launcher extends BotFragment implements Thread.UncaughtExceptionHan
 
         INSTANCE.start();
 
-        BotLog.info("启动 成功 （￣～￣)");
-
     }
 
     @Override
@@ -77,7 +80,7 @@ public class Launcher extends BotFragment implements Thread.UncaughtExceptionHan
         StaticLog.info("正在停止Bot");
 
         INSTANCE.bot().removeGetUpdatesListener();
-        
+
         FollowerTrackTask.stop();
 
 		//  BotServer.INSTACNCE.stop();
@@ -105,7 +108,7 @@ public class Launcher extends BotFragment implements Thread.UncaughtExceptionHan
 		} catch (IOException e) {}
 
 	}
-	
+
 	public void reboot() {
 
 		try {
@@ -115,25 +118,81 @@ public class Launcher extends BotFragment implements Thread.UncaughtExceptionHan
 		} catch (IOException e) {}
 
 	}
-	
-	
+
+    String initMsg = "BOT正在初始化... 这可能需要几分钟的时间.";
+
 	@Override
 	public boolean onMsg(UserData user,Msg msg) {
 
+        if (initIng.get() && msg.isCommand()) {
+
+            msg.send(initMsg).exec();
+
+            return true;
+
+        }
+
+        if (super.onMsg(user,msg)) return true;
+
 		if (!(Env.FOUNDER.equals(user.userName) && msg.isCommand())) return false;
-		
+
 		switch (msg.command()) {
-			
+
 			case "stop" : stop();break;
 			case "restart" : restart();break;
 			case "reboot" : reboot();break;
-			
+
 			default : return false;
-			
+
 		}
-		
+
 		return true;
 
 	}
+
+    @Override
+    public boolean onPPM(UserData user,Msg msg) {
+
+        if (initIng.get()) {
+
+            msg.send(initMsg).exec();
+
+            return true;
+
+        }
+
+        return super.onPPM(user,msg);
+
+    }
+
+    @Override
+    public boolean onCallback(UserData user,Callback callback) {
+
+        if (initIng.get()) {
+
+            callback.alert(initMsg);
+
+            return true;
+
+        }
+
+        return super.onCallback(user,callback);
+
+    }
+
+    @Override
+    public boolean onQuery(UserData user,Query inlineQuery) {
+
+        if (initIng.get()) {
+
+            inlineQuery.article("请稍后再使用...",initMsg).reply();
+
+            return true;
+
+        }
+
+        return super.onQuery(user,inlineQuery);
+
+    }
 
 }
